refactor(satellite): Now using `AssetCollection` for sprites and config
