+++ src/entities/satellite/grabber.rs
@@ -5,7 +5,7 @@ use bevy_ggrs::{GgrsSchedule, LocalPlayers};
 use leafwing_input_manager::prelude::ActionState;
 
 use super::SatelliteSet;
-use super::{SatelliteConfig, SatelliteConfigHandle};
+use super::assets::{SatelliteAssets, SatelliteConfig};
 use crate::core::inputs::{PlayerAction, PlayerActionState};
 use crate::core::physics::{Position, Velocity};
 use crate::entities::player::Player;
@@ -76,10 +76,10 @@ fn detect_player_entry(
     mut commands: Commands,
     player_query: Query<(Entity, &Position), With<Player>>,
     grabber_query: Query<(Entity, &Position), With<Grabber>>,
-    config_handle: Res<SatelliteConfigHandle>,
     configs: Res<Assets<SatelliteConfig>>,
+    assets: Res<SatelliteAssets>,
 ) {
-    let Some(config) = configs.get(&config_handle.0) else {
+    let Some(config) = configs.get(&assets.config) else {
         warn!("Satellite config not loaded yet");
         return;
     };
