+++ src/entities/satellite/mod.rs
@@ -2,13 +2,14 @@ use crate::core::physics::{PhysicsSet, Position};
 use bevy::prelude::*;
 use bevy::render::mesh::Mesh;
 use bevy::sprite::Material2dPlugin;
-use bevy_common_assets::ron::RonAssetPlugin;
 
+pub mod assets;
 pub mod bumper;
 pub mod grabber;
 pub mod graviton;
 mod visuals;
 
+use assets::{SatelliteAssets, SatelliteConfig};
 use bevy_ggrs::GgrsSchedule;
 use bumper::Bumper;
 use grabber::Grabber;
@@ -33,21 +34,6 @@ pub struct SpawnSatelliteEvent {
     pub kind: SatelliteKind,
 }
 
-#[derive(serde::Deserialize, Asset, TypePath)]
-pub struct SatelliteConfig {
-    pub orbit_radius: f32,
-    pub min_angular_speed: f32,
-    pub orbit_duration: f32,
-    pub orbit_cooldown: f32,
-    pub decay_rate: f32,
-    pub bump_radius: f32,
-    pub bump_multiplier: f32,
-    pub grabber_radius: f32,
-}
-
-#[derive(Resource)]
-pub struct SatelliteConfigHandle(pub Handle<SatelliteConfig>);
-
 #[derive(SystemSet, Debug, Clone, PartialEq, Eq, Hash)]
 enum SatelliteSet {
     Graviton,
@@ -59,7 +45,6 @@ pub struct SatellitePlugin;
 impl Plugin for SatellitePlugin {
     fn build(&self, app: &mut App) {
         app.register_type::<graviton::Orbited>()
-            .add_plugins(RonAssetPlugin::<SatelliteConfig>::new(&[]))
             .add_plugins(Material2dPlugin::<OrbitMaterial>::default())
             .configure_sets(
                 GgrsSchedule,
@@ -72,37 +57,28 @@ impl Plugin for SatellitePlugin {
                     .in_set(PhysicsSet::Interaction),
             )
             .add_event::<SpawnSatelliteEvent>()
-            .add_systems(Startup, load_satellite_config)
-            .add_systems(Update, handle_spawn_satellite)
+            .add_systems(
+                Update,
+                handle_spawn_satellite.run_if(resource_exists::<SatelliteAssets>),
+            )
             .add_plugins(graviton::GravitonPlugin)
             .add_plugins(bumper::BumperPlugin)
             .add_plugins(grabber::GrabberPlugin);
     }
 }
 
-fn load_satellite_config(mut commands: Commands, asset_server: Res<AssetServer>) {
-    let handle = asset_server.load("config/satellites.ron");
-    commands.insert_resource(SatelliteConfigHandle(handle));
-}
-
 fn handle_spawn_satellite(
     mut commands: Commands,
-    asset_server: Res<AssetServer>,
     mut events: EventReader<SpawnSatelliteEvent>,
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<OrbitMaterial>>,
-    config_handle: Res<SatelliteConfigHandle>,
     configs: Res<Assets<SatelliteConfig>>,
+    assets: Res<SatelliteAssets>,
 ) {
-    let Some(config) = configs.get(&config_handle.0) else {
+    let Some(config) = configs.get(&assets.config) else {
         warn!("Satellite config not loaded yet");
         return;
     };
-    let graviton_active = asset_server.load("img/satellites/working_graviton.png");
-    let graviton_inactive = asset_server.load("img/satellites/destroyed_graviton.png");
-
-    let bumper_texture = asset_server.load("img/satellites/working_bumper.png");
-    let grabber_texture = asset_server.load("img/satellites/working_grabber.png");
 
     for event in events.read() {
         let mut entity = commands.spawn((
@@ -118,15 +94,15 @@ fn handle_spawn_satellite(
                 entity.insert((
                     Graviton,
                     GravitonVisual {
-                        active: graviton_active.clone(),
-                        inactive: graviton_inactive.clone(),
+                        active: assets.working_graviton.clone(),
+                        inactive: assets.destroyed_graviton.clone(),
                     },
                 ));
 
                 entity.with_children(|parent| {
                     parent.spawn((
                         Sprite {
-                            image: graviton_active.clone(),
+                            image: assets.working_graviton.clone(),
                             ..default()
                         },
                         child_transform,
@@ -138,7 +114,7 @@ fn handle_spawn_satellite(
                 entity.with_children(|parent| {
                     parent.spawn((
                         Sprite {
-                            image: bumper_texture.clone(),
+                            image: assets.working_bumper.clone(),
                             ..default()
                         },
                         child_transform,
@@ -150,7 +126,7 @@ fn handle_spawn_satellite(
                 entity.with_children(|parent| {
                     parent.spawn((
                         Sprite {
-                            image: grabber_texture.clone(),
+                            image: assets.working_grabber.clone(),
                             ..default()
                         },
                         child_transform,
